core/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
fastimage==2.0.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
outcome==1.2.0
Pillow==9.5.0
pycparser==2.21
PySocks==1.7.1
pytz==2023.3
selenium==4.9.0
Selenium-Screenshot==2.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.15
wsproto==1.2.0
yarl==1.9.2
